markings: Fix all the remaining license and copyright markings as I can
Fix all the remaining license and copyright markings as I can.

There are still the files in docs/doxygen/doxyxml/generated/ that are
not of easily tracable origin.  They are computer generated files from
*xsd input files that are no longer around.  Technically the files are
owned/copyright by the owner of the *.xsd files which generateDS.py
operated on to generate the files.  See the comments in those files
for where the trail goes cold.