lib: Properly handle out of input for samples/symbol > 8 in M&M clock recovery
The M&M clock recovery block, when the samples/symbol (a.k.a. omega) was
greater than 8, could errantly consume() more input than was provided.
This change fixes that, by allowing the main loop to roll back the
state to the previous iteration.  This change also tries to ensure at
least one output sample will still be emitted, even if the main loop
has to roll back a computed output.

The changes that allow rolling back the state will be useful for
correcting tag propagation later.