lib: Fix error in application of timing estimate in clock_recovery_mm_ff
The PLL in clock_recovery_mm_ff was incorrectly applying the new
timing estimate, the unwrapped d_mu, by aligning the estimated
clock period in time with the current input sample time, as opposed
to aligning it with the current interpolated output sample time.
This change fixes that.

As a consequence it now becomes obvious that d_omega is tracking
the nominal clock period minus half a sample.  The half a sample time,
0.5, is the expected value of the wrapped d_mu, since the wrapped d_mu
ranges in [0.0f, 1.0f] instead of [-0.5f, 0.5f].  This fix keeps
the relation between d_omega, d_mu, and the nominal samples/symbol
the same internally, but externally omega is presented to the user
as the nominal clock period in samples/symbol, and not as the
nominal clock period minus half a sample.