examples: In nwr_fg.grc, allow underdamples clock rec and avoid segfault
Modify the flowgraph variables to allow the M&M clock recovery to easily
be operated in an underdamped mode as well.

Tweak the clock recovery default loop bandwidth (0.04) and damping factor (2.0)
yet again.

Avoid using tag triggering in the QT GUI Constellation sink as it has a
bug that causes a SIGSEGV.