symbol_sync: Switch to the wrapped interpolating_resampler class
Convert the symbol sync blocks internally to use the wrapped
interpolating resampler class to keep track of interpolator phase
and to perform sample interpolation.

Right now, we still have the interpolator fixed to the 8-tap
MMSE FIR interpolator and differentiator.  Upcoming changes
will allow the interpolator to be user selectable.